+++ transports/changelog.md
@@ -2,4 +2,6 @@
 <!-- Old changelogs are automatically attached to the GitHub releases -->
 
 - chore: version update framework to 1.1.18 and core to 1.2.16
-- Adds env variable support for postgres config
+- feat: Use all keys for list models request
+- fix: handled panic when using gemini models with openai integration responses API requests
+- chore: Added id, object, and model fields to Chat Completion responses from Bedrock and Cohere providers